% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{bathymetry}
\alias{bathymetry}
\title{A stylized bathymetric data set of the Sea of Cortez and adjacent areas.}
\format{
An sf tibble, with a single attribute giving elevation in meters (depth is stored as negative elevation).
\describe{
\item{elevation}{Elevation of the contour line, in meters -- negative values below sea level.}
\item{geometry}{An sf linestring -- a bathymetric contour.}
}
}
\usage{
bathymetry
}
\description{
A stylized dataset of bathymetric contours, intended for mapping (not analytical) applications.
The source raster was aggregated, contoured, filtered for small pieces, simplified, corner-cut, and simplified again.
The contours were generated from GEBCO Compilation Group (2024) GEBCO 2024 Grid (doi:10.5285/1c44ce99-0a0d-5f4f-e063-7086abc0ea0f).
}
\examples{
data(bathymetry)

library(ggplot2)
ggplot() +
 geom_sf(data = bathymetry, aes(color = elevation)) +
 theme_minimal()
}
\keyword{datasets}
