% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{land}
\alias{land}
\title{Land and islands of Western Mexico}
\format{
An sf tibble, with one multipolygon per country in the area.
\describe{
\item{name}{Name of the country.}
\item{geometry}{A simplified multipolygon geometry.}
}
}
\usage{
land
}
\description{
A simplified data set of land polygons and their associated nations, adapted from Natural Earth's global products.
Built by combining the high-resolution land data set with the minor islands data set.
Islands that lacked administrative info in Natural Earth were assigned to their nearest nation.
Based on visual inspection the match is good; any mismatch is not intentional -- \emph{viva la Mexico} and god bless the USA.
"Made with Natural Earth. Free vector and raster map data @ naturalearthdata.com."
}
\examples{
data(land)

library(ggplot2)
ggplot() +
 geom_sf(data = land, aes(fill = name)) +
 theme_minimal()
}
\keyword{datasets}
