% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{landcover}
\alias{landcover}
\title{Major landcover classes of Western Mexico}
\format{
An sf tibble with one row per landcover polygon.
\describe{
\item{class}{Name of the landcover class (slightly renamed from the source in a few cases).}
\item{geometry}{A simplified polygon geometry.}
}
}
\usage{
landcover
}
\description{
A simplified data set of landcover types, derived from the global EarthEnv product.
Tuanmu, M.-N. and W. Jetz. 2014. A global 1-km consensus land-cover product for biodiversity and ecosystem modeling. Global Ecology and Biogeography 23(9): 1031-1045.
}
\examples{
data(landcover)

library(ggplot2)
ggplot() +
 geom_sf(data = landcover, aes(fill = class)) +
 theme_minimal()
}
\keyword{datasets}
