% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{lc_pal}
\alias{lc_pal}
\title{Color palette for landcover data}
\format{
A named character vector of hex color codes, with names matching the levels of \code{landcover$class}.
}
\usage{
lc_pal
}
\description{
A hand-picked set of hex colors (via coolors.co) intended to pair reasonably with the \code{landcover} classes.
}
\examples{
data(lc_pal)
data(landcover)

library(ggplot2)
ggplot() +
geom_sf(data = landcover, aes(fill = class)) +
 scale_fill_manual(values = lc_pal) +
 theme_minimal()
}
\keyword{datasets}
