% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{protected}
\alias{protected}
\title{Protected areas of Mexico}
\format{
An sf tibble, with a handful of attributes briefly describing each protected area.
\describe{
\item{name}{Official name of the protected area.}
\item{date_established}{Date the preserve was established.}
\item{reserve_type}{A locally defined category assigned to each reserve. \code{Marine} and \code{Terrestrial} are reserved for areas with 100\\% of their surface in that category; \verb{Primarily Marine} and \verb{Primarily Terrestrial} are used for mixed areas, based on which surface type dominates.}
\item{geometry}{A simplified multipolygon geometry.}
}
}
\usage{
protected
}
\description{
A very lightly simplified data set of protected areas in Mexico, for cartographic purposes.
From CONABIO CONANP, (09/2024). 'Federal Protected Natural Areas of Mexico, September 2024', National Commission for the Knowledge and Use of Biodiversity. Mexico City, Mexico.
}
\examples{
data(protected)

library(ggplot2)
ggplot() +
 geom_sf(data = protected, aes(fill = reserve_type)) +
 theme_minimal()
}
\keyword{datasets}
