% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{bathymetry}
\alias{bathymetry}
\title{A stylized bathymetric data set of the Sea of Cortez and adjacent areas.}
\format{
sf/tibble/dataframe, with a single variable detailing the elevation in meters.
\describe{
\item{Depth}{Depth of the contour line in meters.}
\item{geometry}{a sf linestring, bathymetric contours.}
}
}
\usage{
bathymetry
}
\description{
A stylized dataset of bathymetric contours for mapping (not analytical) applications.
Data were aggregated, contoured, had small pieces were removed, simplified, corner cut, and simplified again.
The contours were generated from GEBCO Compilation Group (2024) GEBCO 2024 Grid (doi:10.5285/1c44ce99-0a0d-5f4f- e063-7086abc0ea0f).
}
\examples{
data(bathymetry)

library(ggplot2)
ggplot() +
 geom_sf(data = bathymetry, aes(color = Elevation))
}
\keyword{datasets}
