% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{land}
\alias{land}
\title{Land and islands of Western Mexico}
\format{
sf/tibble/dataframe, with two multipolygons, one for each of the countries in the area.
\describe{
\item{Name}{Name of the country.}
\item{geometry}{A simplified multipolygon geometry.}
}
}
\usage{
land
}
\description{
A simplified data set of land, and nations, (for global usage) from Natural Earth.
Basically, combined the high res land data set with the minor islands data set.
Islands, which do not contain administrative info from Natural Earth were assigned by being 'nearest' feature to a nation.
Based on visual inspection the match is good; any mismatch is not my intention - \emph{viva la Mexico} and god bless the USA.
"Made with Natural Earth. Free vector and raster map data @ naturalearthdata.com."
}
\examples{
data(land)

library(ggplot2)
ggplot() +
 geom_sf(data = land, aes(fill = name))
}
\keyword{datasets}
