% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{landcover}
\alias{landcover}
\title{Major landcover classes of Western Mexico}
\format{
sf/tibble/dataframe, with a handful of variables briefly describing each locality.
\describe{
\item{class}{Name of the class, slightly modified in several instances. }
\item{geometry}{A simplified polygon geometry.}
}
}
\usage{
landcover
}
\description{
A simplified data set of landcover types (for global usage) from EarthEnv.
Tuanmu, M.-N. and W. Jetz. 2014. A global 1-km consensus land-cover product for biodiversity and ecosystem modeling. Global Ecology and Biogeography 23(9): 1031-1045.
}
\examples{
data(landcover)

library(ggplot2)
ggplot() +
 geom_sf(data = landcover, aes(fill = class))
}
\keyword{datasets}
