% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{lc_pal}
\alias{lc_pal}
\title{Color palette for landcover data}
\format{
An object of class \code{character} of length 12.
}
\usage{
lc_pal
}
\description{
Colors from coolors, hand selected by me, to look OK.
}
\examples{
data(lc_pal)

library(ggplot2)
ggplot() +
geom_sf(data = landcover, aes(fill = class)) +
 scale_fill_manual(values = lc_pal) +
 theme_minimal()
}
\keyword{datasets}
