% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{places}
\alias{places}
\title{Places which the Western Flyer passed through on Ricketts and Steinbecks journey to the sea of Cortez.}
\format{
sf/tibble/dataframe with two date columns.
\describe{
\item{location-espanol}{A currently used place name in spanish.}
\item{location-english}{The name of the location used in the book.}
\item{collect}{Boolean. Whether the crew collected specimens there.}
\item{date_arrive}{The date the Western flyer arrived (or passed the place).}
\item{date_depart}{The date the Western flyer departed, or passed the place.}
\item{geometry}{The geometry of the place, not 100\% accurate, and often slightly off shore.}
}
}
\usage{
places
}
\description{
A dataset with place names provided both following the text, and more modern Mexican place names in Spanish.
Note that several of the book place names seem somewhat peculiar, but the localities seem well matched to the current names in spanish.
}
\keyword{datasets}
