% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{route}
\alias{route}
\title{A stylized route of the Western Flyer on Ricketts and Steinbecks journey to the sea of Cortez.}
\format{
sf/tibble/dataframe, with a single variable to join to the 'places' data set (if desired).
\describe{
\item{Destination}{The name of locations used in the book. Can be used to join to the "places" dataset to get dates of travel.}
\item{geometry}{a sf linestring, from the previous destination to the current one.}
}
}
\usage{
route
}
\description{
A dataset with a destination and a route to that destination.
Note that the route is fictionalized from the book, and modified to be aesthetically pleasing, rather than nautically accurate.
}
\examples{
data(route)
plot(route[,'date_arrive'])

places <- st_drop_geometry(places)
route <- left_join(route, st_drop_geometry(places), by = c('Destination' = 'location.english'))

plot(route[,6])

library(ggplot2)
ggplot() + 
 geom_sf(data = route, aes(color = date_arrive))

}
\keyword{datasets}
